Added hosting extension
