+++ Src/SM-Post/Post.Cmd/Post.Cmd.Api/Program.cs
@@ -1,54 +1,16 @@
 using Confluent.Kafka;
-using CQRS.Core.Domain;
-using CQRS.Core.Handlers;
-using CQRS.Core.Infrastructure;
-using CQRS.Core.Producers;
-using Post.Cmd.Api.Commands;
-using Post.Cmd.Domain.Aggregates;
+
+using Post.Cmd.Api.Extensions.HostingExtensions;
 using Post.Cmd.Infrastructure.Config;
-using Post.Cmd.Infrastructure.Dispatchers;
-using Post.Cmd.Infrastructure.Handlers;
-using Post.Cmd.Infrastructure.Producers;
-using Post.Cmd.Infrastructure.Repositories;
-using Post.Cmd.Infrastructure.Stores;
 
 var builder = WebApplication.CreateBuilder(args);
 
-
-builder.Services.AddControllers();
-
 builder.Services.Configure<MongoConfig>(builder.Configuration.GetSection(nameof(MongoConfig)));
 builder.Services.Configure<ProducerConfig>(builder.Configuration.GetSection(nameof(ProducerConfig)));
-builder.Services.AddScoped<IEventStoreRepository, EventStoreRepository>();
-builder.Services.AddScoped<IEventStore, EventStore>();
-builder.Services.AddScoped<IEventProducer, EventProducer>();
-builder.Services.AddScoped<IEventSourcingHandler<PostAggregate>, EventSourcingHandler>();
-builder.Services.AddScoped<ICommandHandler, CommandHandler>();
 
-builder.Services.AddEndpointsApiExplorer();
-builder.Services.AddSwaggerGen();
-
-// register dispatcher handlers
-builder.Services.AddSingleton<ICommandDispatcher>(sp =>
-{
-    var scope = sp.CreateScope();
-    var commandHandler = scope.ServiceProvider.GetRequiredService<ICommandHandler>();
-
-    var dispatcher = new CommandDispatcher();
-
-    dispatcher.RegisterHandler<NewPostCommand>(commandHandler.HandleAsync);
-    dispatcher.RegisterHandler<EditMessageCommand>(commandHandler.HandleAsync);
-    dispatcher.RegisterHandler<AddCommentCommand>(commandHandler.HandleAsync);
-    dispatcher.RegisterHandler<EditCommentCommand>(commandHandler.HandleAsync);
-    dispatcher.RegisterHandler<LikePostCommand>(commandHandler.HandleAsync);
-    dispatcher.RegisterHandler<RemoveCommentCommand>(commandHandler.HandleAsync);
-    dispatcher.RegisterHandler<DeletePostCommand>(commandHandler.HandleAsync);
-
-    return dispatcher;
-});
+builder.Services.AddServices();
 
 var app = builder.Build();
-
 
 if (app.Environment.IsDevelopment())
 {
